+++ examples/complete/main.tf
@@ -10,6 +10,21 @@ module "comeplete_endpoints_example" {
       private_dns_enabled = var.private_dns_enabled
       subnet_ids          = local.internal_subnets
       name                = "ssm"
+      auto_accept         = true
+      ip_address_type     = "ipv4"
+
+      dns_options = [
+        {
+          dns_record_ip_type                             = "ipv4"
+          private_dns_only_for_inbound_resolver_endpoint = false
+        }
+      ]
+
+      timeouts = {
+        create = "15m"
+        update = "15m"
+        delete = "15m"
+      }
     },
     {
       service_name        = "${local.reverse_dns_prefix}.${local.region}.ec2messages"